hood pivot command
