+++ src/main/java/frc/robot/RobotContainer.java
@@ -19,8 +19,8 @@ import frc.robot.commands.IntakeOutCommand;
 import frc.robot.commands.IntakeInCommand;
 import frc.robot.commands.IntakeIndexCommand;
 import frc.robot.commands.OuttakeIndexCommand;
-import frc.robot.commands.PivotBackwardCommand;
-import frc.robot.commands.PivotForwardCommand;
+import frc.robot.commands.HoodPivotBackwardCommand;
+import frc.robot.commands.HoodPivotForwardCommand;
 import frc.robot.commands.DistanceDriveCommand;
 import frc.robot.commands.RotationCommand;
 import frc.robot.commands.ShooterAmpCommand;
