+++ src/main/java/frc/robot/commands/HoodEncoderTest.Java
@@ -8,32 +8,29 @@ import frc.robot.Constants.HoodConstants;
 /**
  * Puts hood backwards (in speaker shooting position).
  */
-public class PivotBackwardCommand extends Command {
+public class HoodEncoderTest extends Command {
     private double m_encoderStart;
     private Hood m_hood;
 
-    public PivotBackwardCommand(Hood hood){
+    public HoodEncoderTest(Hood hood){
         m_hood = hood;
     }
 
     @Override
     public void initialize() {
-        m_encoderStart = m_hood.getPivotEncoderPosition();
     }
 
     @Override
     public void execute() {
-        m_hood.runPivotHoodMotor(-HoodConstants.PIVOT_MOTOR_SPEED);
+      System.out.println(m_hood.getPivotEncoderPosition());
     }
 
     @Override
     public void end(boolean interrupted) {
-        m_hood.pivotMotorOff();
     }
 
     @Override
     public boolean isFinished() {
-        return m_hood.getPivotEncoderPosition() <= m_encoderStart - HoodConstants.PIVOT_DELTA;
-        // return m_hood.getPivotEncoderPosition() <= IntakeConstants.PIVOT_UP_POSITION;
+        return false;
     }
 }
